# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus Pro Backend API
  version: 0.2.0
  description: |
    Identifiers for persisted tenant resources are globally unique across tenants.
    tenant_id scopes ownership, authorization, and row-level security.
security:
  - bearerAuth: []
paths:
...
  /v1/hr/positions:
    get:
      summary: List positions
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: Position page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionListDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
    post:
      summary: Create position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: Created position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/hr/employees:
    get:
      summary: List employees
      parameters:
        - {name: keyword, in: query, schema: {type: string}}
        - {name: department_id, in: query, schema: {type: string}}
        - {name: employment_status, in: query, schema: {type: string}}
        - {name: category, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
        - {name: sort, in: query, schema: {type: string, enum: [created_at_asc, created_at_desc, hire_date_asc, hire_date_desc]}}
      responses:
        "200":
          description: Employee page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
    post:
      summary: Create employee aggregate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetailDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/hr/employee-options:
    get:
      summary: Employee form options
      responses:
        "200":
          description: Employee options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeOptionsDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/ehrms/sync:
    post:
      summary: Start a tracked manual eHRMS sync
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/StartEHRMSSyncRequest"}
      responses:
        "200":
          description: Completed tracked sync run
          content:
            application/json:
              schema: {$ref: "#/components/schemas/EHRMSSyncRunDetailDataResponse"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/org/units:
    get:
      summary: List organization units
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: Organization unit page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
    post:
      summary: Create organization unit
      responses:
        "201":
          description: Created organization unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
  /v1/org/units/ehrms/sync:
    post:
      summary: Sync organization units from configured eHRMS
      responses:
        "200":
          description: eHRMS organization unit sync summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EHRMSOrgUnitSyncDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/org/units/{id}:
    patch:
      summary: Update organization unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code: {type: string}
                name: {type: string}
                parent_id: {type: string}
                manager_position_id: {type: string}
                closed:
                  type: boolean
                  description: Closing an organization unit also closes all descendants; a child cannot reopen while an ancestor remains closed.
      responses:
        "200":
          description: Updated organization unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token issued by the configured identity provider and mapped to a local tenant account.
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
    Sort:
      name: sort
      in: query
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Validation failed. Field errors include tab, field, code and message; import row errors include row_number.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: Authenticated tenant/account context is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Permission, menu, button, field, or data-scope access was denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource was not found in the current tenant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error. Response includes trace_id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    DataResponse:
      type: object
      required: [data]
      properties:
        data:
          description: Successful response payload.
...
    EmployeeDetailDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EmployeeDetail"
    EmployeeListDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EmployeeListResponse"
...
    EmployeeOptionsDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EmployeeOptions"
...
    EHRMSSyncRunDetailDataResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/EHRMSSyncRunDetail"}
    EHRMSOrgUnitSyncDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EHRMSOrgUnitSyncResponse"
...
    PositionDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Position"
    PositionListDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PositionListResponse"
...
    PositionInput:
      type: object
      required: [name]
      properties:
        code: {type: string}
        name: {type: string}
        org_unit_id:
          type: string
          description: Optional tenant-local org unit id. Empty means position is not bound to one org unit.
        level: {type: string}
        status:
          $ref: "#/components/schemas/PositionStatus"
        description: {type: string}
...
    PositionStatus:
      type: string
      enum: [active, disabled]
    PositionListResponse:
      type: object
      required: [items, total, page, page_size, sort]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Position"}
        total: {type: integer}
        page: {type: integer}
        page_size: {type: integer}
        sort: {type: string}
...
    EmployeeInput:
      type: object
      required: [name, company_email, org_unit_id, position, category, hire_date, basic_info, employment_info, education_military_info, contact_info, insurance_info]
      properties:
        employee_no:
          type: string
          description: Optional. Generated by the backend with IKL prefix when omitted.
        name: {type: string}
        company_email: {type: string, format: email}
        personal_email: {type: string, format: email}
        phone: {type: string}
        org_unit_id: {type: string}
        account_id: {type: string}
        account_policy:
          type: string
          enum: [none, link_existing, create_pending_invite, create_active]
          description: Optional account lifecycle policy. Defaults to link_existing when account_id is provided, otherwise none.
        position_id:
          type: string
          description: Optional position entity id. When supplied, employee.position is synchronized from the position name.
        position: {type: string}
        category:
          $ref: "#/components/schemas/EmployeeCategory"
        status:
          $ref: "#/components/schemas/EmployeeStatusWithDeleted"
        employment_status:
          $ref: "#/components/schemas/EmployeeStatusWithDeleted"
        hire_date: {type: string}
        resign_date: {type: string}
        basic_info: {$ref: "#/components/schemas/EmployeeBasicInfo"}
        employment_info: {$ref: "#/components/schemas/EmployeeEmploymentInfo"}
        education_military_info: {$ref: "#/components/schemas/EmployeeEducationMilitaryInfo"}
        contact_info: {$ref: "#/components/schemas/EmployeeContactInfo"}
        insurance_info: {$ref: "#/components/schemas/EmployeeInsuranceInfo"}
        internal_experiences:
          type: array
          items: {$ref: "#/components/schemas/EmployeeExperience"}
...
    EmployeeListResponse:
      type: object
      required: [items, total, page, page_size, sort]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Employee"}
        total: {type: integer}
        page: {type: integer}
        page_size: {type: integer}
        sort: {type: string}
...
    StartEHRMSSyncRequest:
      type: object
      properties:
        mode: {type: string, enum: [create, update, upsert]}
        since: {type: string, format: date}
        include_attendance: {type: boolean, default: false}
...